Fix: 게시판 댓글 컴포넌트 반복렌더링 수정
fetchComments 함수를 단한번만 구현되도록
